+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/fivetran/go-fivetran/compare/v0.9.2...HEAD)
+## [Unreleased](https://github.com/fivetran/go-fivetran/compare/v0.9.3...HEAD)
+
+## [0.9.3](https://github.com/fivetran/go-fivetran/compare/v0.9.2...v0.9.3)
+
+Supported the following Fivetran API endpoints:
+- Transformations for dbt Core Management: [Test dbt Project](https://fivetran.com/docs/rest-api/transformations-for-dbt-core-management#testdbtproject)
+
+## [0.9.2](https://github.com/fivetran/go-fivetran/compare/v0.9.1...v0.9.2)
+
+## Fixed
+- Issue with `DestinationConfigResponse.data.config.is_private_key_encrypted`
 
 ## [0.9.1](https://github.com/fivetran/go-fivetran/compare/v0.9.0...v0.9.1)